safekeeper: rename wal_backup and wal_backup_partial files to wal_upload and wal_upload_partial
