+++ safekeeper/src/lib.rs
@@ -37,8 +37,8 @@ pub mod timeline_eviction;
 pub mod timeline_guard;
 pub mod timeline_manager;
 pub mod timelines_set;
-pub mod wal_backup;
-pub mod wal_backup_partial;
+pub mod wal_upload;
+pub mod wal_upload_partial;
 pub mod wal_reader_stream;
 pub mod wal_service;
 pub mod wal_storage;
